Move schemes for travels
